+++ common/src/iosMain/kotlin/tk/zwander/common/util/UrlHandler.ios.kt
@@ -3,14 +3,20 @@ package tk.zwander.common.util
 import io.ktor.http.encodeURLQueryComponent
 import platform.Foundation.NSURL
 import platform.UIKit.UIApplication
+import tk.zwander.common.util.BifrostLogger
 
 actual object UrlHandler {
     actual fun launchUrl(url: String, forceBrowser: Boolean) {
-        UIApplication.sharedApplication.openURL(
-            NSURL.URLWithString(url)!!,
-            mapOf<Any?, Any?>(),
-            null,
-        )
+        val nsUrl = NSURL.URLWithString(url)
+        if (nsUrl != null) {
+            UIApplication.sharedApplication.openURL(
+                nsUrl,
+                mapOf<Any?, Any?>(),
+                null,
+            )
+        } else {
+            BifrostLogger.general.warn("launchUrl: invalid URL: $url")
+        }
     }
 
     actual fun sendEmail(address: String, subject: String?, content: String?) {
@@ -32,10 +38,15 @@ actual object UrlHandler {
             }
         }
 
-        UIApplication.sharedApplication.openURL(
-            NSURL.URLWithString("mailto:$url")!!,
-            mapOf<Any?, Any?>(),
-            null,
-        )
+        val nsUrl = NSURL.URLWithString("mailto:$url")
+        if (nsUrl != null) {
+            UIApplication.sharedApplication.openURL(
+                nsUrl,
+                mapOf<Any?, Any?>(),
+                null,
+            )
+        } else {
+            BifrostLogger.general.warn("sendEmail: invalid mailto URL: $url")
+        }
     }
 }